video: support colors in truetype console
In the UEFI context we use colored output. When printing truetype letters
we have to interpolate the pixel color between the foreground and
background color according to the gray value of the font pixel.

We could speed up the output a by writing a lookup table in the uclass
whenever a color changes but this would impose a burden on all console
drivers. So this is not implemented.

For testing the sandbox can be used when enabling CONFIG_EFI_SELFTEST.

    $ ./u-boot -D -l
    => setenv efi_selftest text output
    => bootefi selftest

This test writes text in all combinations of foreground and background
colors to the console.

To test 16bit mode modify arch/sandbox/dts/sandbox.dtsi. Set

    /lcd/log2-depth = <4>;

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>